Provide PNG badges for reference from HuggingFace
Add PNG images for all the badges so that they can be used in HuggingFace
which fails to display SVG images.
Change the provided markup on the Badges page to refer to the PNG images
and link to the model view.
Limitation: While the SVG bages show the percentage for in progress the PNG
version only reads "In progress".

Signed-off-by: Arnaud Le Hors <lehors@us.ibm.com>